backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.9.0
SQLAlchemy==2.0.32
psycopg[binary,pool]==3.2.1
email-validator
